--- a/src/components/code-line.component.js
+++ b/src/components/code-line.component.js
@@ -1,12 +1,12 @@
 // @flow
 /* eslint-disable no-nested-ternary */
 
-import React from 'react';
+import React, { Component } from 'react';
 import { View, Text, StyleSheet, Dimensions } from 'react-native';
-
+import SyntaxHighlighter from 'react-native-syntax-highlighter';
+import { getLanguage } from 'lowlight';
+import { github as GithubStyle } from 'react-syntax-highlighter/dist/styles';
 import { colors, normalize } from 'config';
-
-type Props = { newChunk: boolean, change: Object };
 
 const styles = StyleSheet.create({
   container: {
@@ -63,44 +63,78 @@ const styles = StyleSheet.create({
   },
 });
 
-export const CodeLine = ({ newChunk, change }: Props) =>
-  <View style={styles.container}>
-    <View style={styles.wrapper}>
-      <View
-        style={[
-          styles.lineNumbers,
-          newChunk && styles.newChunkLineNumbers,
-          change.type === 'add' && styles.addLineNumbers,
-          change.type === 'del' && styles.delLineNumbers,
-        ]}
-      >
-        <Text style={styles.codeLineNumber}>
-          {change.type === 'del'
-            ? change.ln
-            : change.type === 'normal'
-              ? change.ln1
-              : change.type === 'add' ? '' : '...'}
-        </Text>
-        <Text style={styles.codeLineNumber}>
-          {change.type === 'add'
-            ? change.ln
-            : change.type === 'normal'
-              ? change.ln2
-              : change.type === 'del' ? '' : '...'}
-        </Text>
-      </View>
+export class CodeLine extends Component {
+  props: {
+    newChunk: boolean,
+    change: Object,
+    filename: string
+  };
 
-      <View
-        style={[
-          styles.codeLineContainer,
-          newChunk && styles.newChunkLineContainer,
-          change.type === 'add' && styles.addLine,
-          change.type === 'del' && styles.delLine,
-        ]}
-      >
-        <Text style={[styles.codeLine, newChunk && styles.newChunkLine]}>
-          {change.content}
-        </Text>
+  isKnownType(language) {
+    return language && getLanguage(language) && !this.props.newChunk;
+  }
+
+  render() {
+    const { newChunk, change, filename } = this.props;
+    const language = (filename || []).split('.').pop();
+    // SyntaxHighlighter doesn't allow Stylesheet class style
+    const customStyle = {
+      backgroundColor: 'transparent',
+      padding: 0
+    };
+
+    return (
+      <View style={styles.container}>
+        <View style={styles.wrapper}>
+          <View
+            style={[
+              styles.lineNumbers,
+              newChunk && styles.newChunkLineNumbers,
+              change.type === 'add' && styles.addLineNumbers,
+              change.type === 'del' && styles.delLineNumbers,
+            ]}
+          >
+            <Text style={styles.codeLineNumber}>
+              {change.type === 'del'
+                ? change.ln
+                : change.type === 'normal'
+                    ? change.ln1
+                    : change.type === 'add' ? '' : '...'}
+            </Text>
+            <Text style={styles.codeLineNumber}>
+              {change.type === 'add'
+                ? change.ln
+                : change.type === 'normal'
+                    ? change.ln2
+                    : change.type === 'del' ? '' : '...'}
+            </Text>
+          </View>
+
+          <View
+            style={[
+              styles.codeLineContainer,
+              newChunk && styles.newChunkLineContainer,
+              change.type === 'add' && styles.addLine,
+              change.type === 'del' && styles.delLine,
+            ]}
+          >
+            {(newChunk || !this.isKnownType(language)) &&
+              <Text style={[styles.codeLine, newChunk && styles.newChunkLine]}>
+                {change.content}
+              </Text>
+            }
+
+            {this.isKnownType(language) &&
+              <SyntaxHighlighter
+                language={language}
+                style={GithubStyle}
+                CodeTag={Text}
+                codeTagProps={{style: styles.codeLine}}
+                customStyle={customStyle}>{change.content}</SyntaxHighlighter>
+            }
+          </View>
+        </View>
       </View>
-    </View>
-  </View>;
+    );
+  }
+}