> Will be great to have syntax highlighting on `PullDiffScreen` for pull requests

As @housseindjirdeh said here : https://github.com/gitpoint/git-point/pull/60#issuecomment-315512222 

So this PR is about this integration.  
Screenshots:
![Screenshots...](https://user-images.githubusercontent.com/9287184/28243667-bbd7e954-69a9-11e7-8796-7cb87b820e6c.png)

Because of how each line are generated (separately), there is a "Syntax Highlighter" on each line. I did some performance tests to show you that it's not very slower than before (but, yes it is! Especially in case of a lot of line diff).

| | Before | Now | |
|:-:|:-:|:-:|:-:|
| PR #49 | 2480 ms | 4090 ms | +64% |
| PR #79 | 700 ms | 1010 ms | +27% |

_Tested on my iPhone 6S and a "manual chronometer" (human imprecision)_