@Antoine38660 2017-07-16
Once you merge #81 and once I rebase `master` CI should pass 😃