@housseindjirdeh 2017-07-16
This is awesome @Antoine38660 🎉 

Unfortunately the diff screen can already be kind of slow with large diff files since we currently render line by line. PR #49 slows numbers down significantly because we introduced `remote-redux-devtools` there (opened an issue for this #82). Comment out [these lines](https://github.com/gitpoint/git-point/blob/master/root.store.js#L11-L22) and you'll notice more normal speeds (and I assume the difference with including SyntaxHighlighter shouldn't be too bad).